requirements.txt
arrow==1.2.3; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
certifi==2022.9.24; python_version >= "3.8" and python_version < "4" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
cffi==1.15.1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
charset-normalizer==2.1.1; python_version >= "3.8" and python_version < "4" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
cognite-extractor-utils==3.0.2; python_full_version >= "3.8.0" and python_full_version < "4.0.0"
cognite-sdk==4.11.0; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
cryptography==38.0.2; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
dacite==1.6.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
decorator==5.1.1; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.5"
idna==3.4; python_version >= "3.8" and python_version < "4" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
jq==1.3.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and sys_platform == "linux" and python_version >= "3.5" or python_full_version >= "3.8.0" and python_full_version < "4.0.0" and sys_platform == "macos" and python_version >= "3.5"
more-itertools==8.14.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.5"
msal==1.20.0; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
oauthlib==3.2.2; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
prometheus-client==0.15.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
psutil==5.9.3; python_full_version >= "3.8.0" and python_full_version < "4.0.0"
pycparser==2.21; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
pyjwt==2.6.0; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
python-dateutil==2.8.2; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
python-dotenv==0.19.2; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.5"
pyyaml==6.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
requests-oauthlib==1.3.1; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
requests==2.28.1; python_version >= "3.8" and python_version < "4" and python_full_version >= "3.8.0" and python_full_version < "4.0.0"
six==1.16.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.6"
typing-extensions==4.4.0; python_full_version >= "3.8.0" and python_full_version < "4.0.0" and python_version >= "3.7"
urllib3==1.26.12; python_full_version >= "3.8.0" and python_version < "4" and python_version >= "3.8" and python_full_version < "4.0.0"
